HandTracking Aura script
Next part is overall adjusting/tweaking the sizes and position of the aura on each hand, also growing the scale of the size of the aura on the hand, firing the built up spell from each hand along the raycast line. Might even keep the raycast off until a fist is made.  Also adding sfx particles when fired and audio for charge and fire.